Renamed NullReader -> NoopReader and renamed package
