+++ standalone-examples/src/main/java/io/pravega/example/noop/BinaryReader.java
@@ -8,7 +8,7 @@
  *   http://www.apache.org/licenses/LICENSE-2.0
  *
  */
-package io.pravega.example.binary;
+package io.pravega.example.noop;
 
 import io.pravega.client.ClientFactory;
 import io.pravega.client.admin.ReaderGroupManager;
